+++ src/main/java/tech/grasshopper/excel/report/cell/ValueOption.java
@@ -1,0 +1,5 @@
+package tech.grasshopper.excel.report.cell;
+
+public enum ValueOption {
+	VALUE, POSITIVE_NUMBER, STATUS_TEXT, STATUS_COLOR
+}