+++ src/main/java/tech/grasshopper/excel/report/table/FeatureScenarioFailSkipTable.java
@@ -1,7 +1,9 @@
 package tech.grasshopper.excel.report.table;
 
-import static tech.grasshopper.excel.report.cell.CellValueOptions.STATUS_BOLD_CELL_OPTIONS;
-import static tech.grasshopper.excel.report.cell.CellValueOptions.getStatusColorCellValueOption;
+import static tech.grasshopper.excel.report.cell.CellStyles.STATUS_TEXT_BOLD_CELL_STYLE;
+import static tech.grasshopper.excel.report.cell.CellStyles.getStatusColorStyle;
+import static tech.grasshopper.excel.report.cell.ValueOption.STATUS_COLOR;
+import static tech.grasshopper.excel.report.cell.ValueOption.STATUS_TEXT;
 
 import java.util.Arrays;
 import java.util.List;
@@ -49,14 +51,14 @@ public class FeatureScenarioFailSkipTable {
 
 			cellOperations.mergeRows(currentRow, (int) feature.getTotalScenarios(), currentCol, columnCellCount[0]);
 			cellOperations.writeValue(new CellReference(currentRow, currentCol), feature.getName(),
-					getStatusColorCellValueOption(feature.getStatus()));
+					getStatusColorStyle(feature.getStatus()), STATUS_COLOR);
 
 			// Move to feature status column
 			currentCol = currentCol + columnCellCount[0];
 
 			cellOperations.mergeRows(currentRow, (int) feature.getTotalScenarios(), currentCol, columnCellCount[1]);
 			cellOperations.writeValue(new CellReference(currentRow, currentCol), feature.getStatus().toString(),
-					STATUS_BOLD_CELL_OPTIONS);
+					STATUS_TEXT_BOLD_CELL_STYLE, STATUS_TEXT);
 
 			// Move to scenario name column
 			currentCol = currentCol + columnCellCount[1];
@@ -65,14 +67,14 @@ public class FeatureScenarioFailSkipTable {
 
 				cellOperations.mergeRows(currentRow, 1, currentCol, columnCellCount[2]);
 				cellOperations.writeValue(new CellReference(currentRow, currentCol), scenario.getName(),
-						getStatusColorCellValueOption(scenario.getStatus()));
+						getStatusColorStyle(scenario.getStatus()), STATUS_COLOR);
 
 				// Move to scenario status column
 				currentCol = currentCol + columnCellCount[2];
 
 				cellOperations.mergeRows(currentRow, 1, currentCol, columnCellCount[3]);
 				cellOperations.writeValue(new CellReference(currentRow, currentCol), scenario.getStatus().toString(),
-						STATUS_BOLD_CELL_OPTIONS);
+						STATUS_TEXT_BOLD_CELL_STYLE, STATUS_TEXT);
 
 				// Move BACK to scenario name column
 				currentCol = currentCol - columnCellCount[2];