+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.4.0 (2023-08-10)
+
+* Cell style reuse
+* Updated extent report version
+* Updated java version to 11
+
 ## 1.3.0 (2023-02-01)
 
 * Author and device sheet